+++ exporter/opentelemetry-exporter-otlp-proto-http/src/opentelemetry/exporter/otlp/proto/http/metric_exporter/__init__.py
@@ -215,6 +215,7 @@ class OTLPMetricExporter(MetricExporter, OTLPMetricExporterMixin):
         return "metrics"
 
     def force_flush(self, timeout_millis: float = 10_000) -> bool:
+        """Nothing is buffered in this exporter, so this method does nothing."""
         return True