+++ CHANGELOG.md
@@ -6,9 +6,22 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
 - Use BoundedAttributes instead of raw dict to extract attributes from LogRecord and Support dropped_attributes_count in LogRecord ([#3310](https://github.com/open-telemetry/opentelemetry-python/pull/3310))
 - Update the body type in the log
   ([$3343](https://github.com/open-telemetry/opentelemetry-python/pull/3343))
+- Add max_scale option to Exponential Bucket Histogram Aggregation 
+  ([#3323](https://github.com/open-telemetry/opentelemetry-python/pull/3323))
+- Use BoundedAttributes instead of raw dict to extract attributes from LogRecord
+  ([#3310](https://github.com/open-telemetry/opentelemetry-python/pull/3310)) 
+- Support dropped_attributes_count in LogRecord and exporters
+  ([#3351](https://github.com/open-telemetry/opentelemetry-python/pull/3351))
+- Add unit to view instrument selection criteria
+  ([#3341](https://github.com/open-telemetry/opentelemetry-python/pull/3341))
+- Upgrade opentelemetry-proto to 0.20 and regen
+  [#3355](https://github.com/open-telemetry/opentelemetry-python/pull/3355))
+
+
 ## Version 1.18.0/0.39b0 (2023-05-04)
 
 - Select histogram aggregation with an environment variable
@@ -25,6 +38,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   ([#3223](https://github.com/open-telemetry/opentelemetry-python/pull/3223))
 - Add speced out environment variables and arguments for BatchLogRecordProcessor
   ([#3237](https://github.com/open-telemetry/opentelemetry-python/pull/3237))
+
 
 ## Version 1.17.0/0.38b0 (2023-03-22)